chore: rename npm packages to use @jstz-dev org
This commit additionally renames jstz-types to types and jstz-sdk to sdk.
