+++ examples/dashboard/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tezos/jstz-dashboard",
+  "name": "@jstz-dev/dashboard",
   "version": "0.0.0",
   "type": "module",
   "scripts": {
@@ -9,7 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@tezos/jstz-sdk": "^0.0.0",
+    "@jstz-dev/sdk": "^0.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "vite-plugin-top-level-await": "^1.4.1",
