+++ examples/fa2/src/index.ts
@@ -1,4 +1,4 @@
-import { isAddress } from "@tezos/jstz";
+import { isAddress } from "@jstz-dev/jstz";
 
 function isArray<T>(check: (item: T) => item is T, list: unknown): list is T[] {
   return (
