+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tezos/jstz-root",
+  "name": "@jstz-dev/jstz-root",
   "private": true,
   "description": "Associated packages for jstz 👨‍⚖️",
   "authors": "TriliTech Ecosystems and Tools Team <contact@trili.tech>",
@@ -20,10 +20,10 @@
     "docs:preview": "vitepress preview docs"
   },
   "dependencies": {
-    "@tezos/jstz": "^0.0.0",
-    "@tezos/jstz-types": "^0.0.0",
-    "@tezos/jstz-sdk": "^0.0.0",
-    "@tezos/jstz-dashboard": "^0.0.0"
+    "@jstz-dev/jstz": "^0.0.0",
+    "@jstz-dev/types": "^0.0.0",
+    "@jstz-dev/sdk": "^0.0.0",
+    "@jstz-dev/dashboard": "^0.0.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.2.1",
