+++ packages/jstz/tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "../../tsconfig.json",
+  "exclude": ["dist"],
   "compilerOptions": {
-    "types": ["@tezos/jstz-types"]
+    "types": ["@jstz-dev/types"],
+    "declaration": true
   }
 }
